+++ src/sig/ld.ts
@@ -1,5 +1,5 @@
 import { getLogger } from "@logtape/logtape";
-import { decodeBase64 } from "@std/encoding/base64";
+import { decodeBase64, encodeBase64 } from "@std/encoding/base64";
 import { encodeHex } from "@std/encoding/hex";
 import jsonld from "jsonld";
 import {
@@ -7,17 +7,136 @@ import {
   fetchDocumentLoader,
 } from "../runtime/docloader.ts";
 import { Activity, CryptographicKey, Object } from "../vocab/vocab.ts";
-import { fetchKey, type KeyCache } from "./key.ts";
+import { fetchKey, type KeyCache, validateCryptoKey } from "./key.ts";
 
 const logger = getLogger(["fedify", "sig", "ld"]);
 
-interface SignedJsonLd {
-  signature: {
-    type: "RsaSignature2017";
-    creator: string;
-    created: string;
-    signatureValue: string;
+/**
+ * A signature of a JSON-LD document.
+ * @since 1.0.0
+ */
+export interface Signature {
+  "@context"?: "https://w3id.org/identity/v1";
+  type: "RsaSignature2017";
+  id?: string;
+  creator: string;
+  created: string;
+  signatureValue: string;
+}
+
+/**
+ * Attaches a LD signature to the given JSON-LD document.
+ * @param jsonLd The JSON-LD document to attach the signature to.  It is not
+ *               modified.
+ * @param signature The signature to attach.
+ * @returns The JSON-LD document with the attached signature.
+ * @throws {TypeError} If the input document is not a valid JSON-LD document.
+ * @since 1.0.0
+ */
+export function attachSignature(
+  jsonLd: unknown,
+  signature: Signature,
+): { signature: Signature } {
+  if (typeof jsonLd !== "object" || jsonLd == null) {
+    throw new TypeError(
+      "Failed to attach signature; invalid JSON-LD document.",
+    );
+  }
+  return { ...jsonLd, signature };
+}
+
+/**
+ * Options for creating Linked Data Signatures.
+ * @since 1.0.0
+ */
+export interface CreateSignatureOptions {
+  /**
+   * The context loader for loading remote JSON-LD contexts.
+   */
+  contextLoader?: DocumentLoader;
+
+  /**
+   * The time when the signature was created.  If not specified, the current
+   * time will be used.
+   */
+  created?: Temporal.Instant;
+}
+
+/**
+ * Creates a LD signature for the given JSON-LD document.
+ * @param jsonLd The JSON-LD document to sign.
+ * @param privateKey The private key to sign the document.
+ * @param keyId The ID of the public key that corresponds to the private key.
+ * @param options Additional options for creating the signature.
+ *                See also {@link CreateSignatureOptions}.
+ * @return The created signature.
+ * @throws {TypeError} If the private key is invalid or unsupported.
+ * @since 1.0.0
+ */
+export async function createSignature(
+  jsonLd: unknown,
+  privateKey: CryptoKey,
+  keyId: URL,
+  { contextLoader, created }: CreateSignatureOptions = {},
+): Promise<Signature> {
+  validateCryptoKey(privateKey, "private");
+  if (privateKey.algorithm.name !== "RSASSA-PKCS1-v1_5") {
+    throw new TypeError("Unsupported algorithm: " + privateKey.algorithm.name);
+  }
+  const options = {
+    "@context": "https://w3id.org/identity/v1" as const,
+    creator: keyId.href,
+    created: created?.toString() ?? new Date().toISOString(),
   };
+  const optionsHash = await hashJsonLd(options, contextLoader);
+  const docHash = await hashJsonLd(jsonLd, contextLoader);
+  const message = optionsHash + docHash;
+  const encoder = new TextEncoder();
+  const messageBytes = encoder.encode(message);
+  const signature = await crypto.subtle.sign(
+    "RSASSA-PKCS1-v1_5",
+    privateKey,
+    messageBytes,
+  );
+  return {
+    ...options,
+    type: "RsaSignature2017",
+    signatureValue: encodeBase64(signature),
+  };
+}
+
+/**
+ * Options for signing JSON-LD documents.
+ * @since 1.0.0
+ */
+export interface SignJsonLdOptions extends CreateSignatureOptions {
+}
+
+/**
+ * Signs the given JSON-LD document with the private key and returns the signed
+ * JSON-LD document.
+ * @param jsonLd The JSON-LD document to sign.
+ * @param privateKey The private key to sign the document.
+ * @param keyId The key ID to use in the signature.  It will be used by the
+ *              verifier to fetch the corresponding public key.
+ * @param options Additional options for signing the document.
+ *                See also {@link SignJsonLdOptions}.
+ * @returns The signed JSON-LD document.
+ * @throws {TypeError} If the private key is invalid or unsupported.
+ * @since 1.0.0
+ */
+export async function signJsonLd(
+  jsonLd: unknown,
+  privateKey: CryptoKey,
+  keyId: URL,
+  options: SignJsonLdOptions,
+): Promise<{ signature: Signature }> {
+  const signature = await createSignature(jsonLd, privateKey, keyId, options);
+  return attachSignature(jsonLd, signature);
+}
+
+interface SignedJsonLd {
+  signature: Signature;
 }
 
 function hasSignature(jsonLd: unknown): jsonLd is SignedJsonLd {